+++ CHANGELOG.md
@@ -8,7 +8,9 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
-*  `epy`: Add frozen dataclass support for `epy.ContextManager`
+*  `epy`:
+    *   Add frozen dataclass support for `epy.ContextManager`
+    *   Support adhoc import of proto files with hyphen.
 *  `enp`: Add `ArraySpec` support for `grain.python.SharedMemoryArrays`.
 
 ## [1.9.4] - 2024-09-03